update/delete routine, relation exercise/routine

+++ src/routines/routines.service.ts
@@ -1,22 +1,45 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ExerciseRepository } from './../exercises/exercises.repository';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { RoutineRepository } from './routines.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { CreateRoutineDto } from './dto/create-routine.dto';
 import { User } from 'src/auth/user.entity';
 import { Routine } from './routines.entity';
+import { Exercise } from 'src/exercises/exercises.entity';
+import { uniq } from 'lodash';
 
 @Injectable()
 export class RoutinesService {
   constructor(
     @InjectRepository(RoutineRepository)
     private routineRepository: RoutineRepository,
+    @InjectRepository(ExerciseRepository)
+    private exerciseRepository: ExerciseRepository,
   ) {}
 
   public async createRoutine(
     createRoutineDto: CreateRoutineDto,
     user: User,
   ): Promise<Routine> {
-    return await this.routineRepository.createRoutine(createRoutineDto, user);
+    const { exercisesId } = createRoutineDto;
+    const exercises: Exercise[] = [];
+    uniq(exercisesId).forEach(async exerciseId => {
+      const exercise = await this.exerciseRepository.findOne({
+        id: exerciseId,
+      });
+      if (exercise) {
+        exercises.push(exercise);
+      }
+    });
+    return await this.routineRepository.createRoutine(
+      createRoutineDto,
+      exercises,
+      user,
+    );
   }
 
   public async getRoutines(user: User): Promise<Routine[]> {
@@ -28,13 +51,26 @@ export class RoutinesService {
   }
 
   async deleteRoutine(id: number, user: User): Promise<void> {
-    const result = await this.routineRepository.delete({ id, userId: user.id });
-    if (result.affected === 0) {
+    const routine = await this.routineRepository.findOne({
+      id,
+      userId: user.id,
+    });
+    if (routine) {
+      if (routine.active) {
+        throw new UnauthorizedException(`Can't delete an active routine`);
+      } else {
+        this.routineRepository.delete({ id, userId: user.id });
+      }
+    } else {
       throw new NotFoundException(`Task with ID ${id} not found`);
     }
   }
 
   async setActiveRoutine(id: number, user: User): Promise<Routine> {
     return await this.routineRepository.setActiveRoutine(user, id);
+  }
+
+  async getExercise(id: number) {
+    return await this.exerciseRepository.findOne({ id });
   }
 }
